0.8.0

+++ tests/tests.xsh
@@ -68,7 +68,15 @@ if __name__ == '__main__':
     if opt.hosts:
         opt.hosts = opt.hosts.split(',')
 
+    xxh = '/xxh/xxh/xxh'
+    local_xxh_home = '/root/local_xxh_home'
+    xxh_args = ['++local-xxh-home', local_xxh_home]
+
     hosts = {}
+    hosts['local'] = {
+        'xxh_auth': [],
+        'home': '/root'
+    }
     hosts['ubuntu_k'] = {
         'user':'root',
         'home':'/root',
@@ -89,14 +97,14 @@ if __name__ == '__main__':
     rm -rf /root/.ssh/known_hosts
 
     if opt.remove:
-        print('Remove start:/root/.xxh')
-        rm -rf /root/.xxh
+        print(f'Remove start:{local_xxh_home}')
+        rm -rf @(local_xxh_home)
 
-    if not p'/root/.xxh/.xxh/shells'.exists():
+    if not pf'{local_xxh_home}/.xxh/shells'.exists():
         print("Don't forget `pip uninstall xxh-xxh` before tests")
         print('First time of executing tests takes time because of downloading files. Take a gulp of water or a few :)')
 
-    mkdir -p /root/.xxh/.xxh/plugins /root/.xxh/.xxh/shells
+    mkdir -p @(local_xxh_home)/.xxh/plugins @(local_xxh_home)/.xxh/shells
 
     print('Prepare repos (to avoid full update use --skip-repos-update)')
 
@@ -121,16 +129,14 @@ if __name__ == '__main__':
     if opt.shell:
         xxh_shell_repos = {opt.shell: xxh_shell_repos[opt.shell]}
 
-    xxh = '/xxh/xxh/xxh'
-
     if not opt.skip_repos_update:
-        print('Remove /root/.xxh/.xxh/shells /root/.xxh/.xxh/plugins')
-        rm -rf /root/.xxh/.xxh/shells /root/.xxh/.xxh/plugins
+        print(f'Remove {local_xxh_home}/.xxh/shells {local_xxh_home}/.xxh/plugins')
+        rm -rf @(local_xxh_home)/.xxh/shells @(local_xxh_home)/.xxh/plugins
 
     for xxh_shell, install_repos in xxh_shell_repos.items():
         for rtype, repos  in install_repos.items():
             for repo in repos:
-                repo_dir = pf'/root/.xxh/.xxh/{rtype}/{repo}'
+                repo_dir = pf'{local_xxh_home}/.xxh/{rtype}/{repo}'
                 repo_local_path = pf'/xxh/{repo}'
 
                 if repo_dir.exists():
@@ -142,11 +148,20 @@ if __name__ == '__main__':
                     if repo_local_path.exists():
                         print(f'Repo {repo}: replaced from /xxh/{repo}. Do not forget to pull from master!')
                         repo = f'{repo}+path+{repo_local_path}'
-                    @(xxh) +I @(repo)
-
-    xxh_args = []
+                    @(xxh) +I @(repo) ++local-xxh-home @(local_xxh_home)
 
     ssh_opts = ["-o", "StrictHostKeyChecking=accept-new", "-o", "LogLevel=QUIET"]
+
+
+    print('[Local test xonsh xxh plugins]')
+    print(f'Remove /root/.xxh_test')
+    rm -rf /root/.xxh_test
+    check(
+        'Local test xonsh xxh plugins',
+        $(echo @(xxh) local +iff +hh /root/.xxh_test +lh /root/.xxh_test +I xxh-plugin-xonsh-theme-bar +I xxh-plugin-xonsh-pipe-liner +I xxh-plugin-xonsh-autojump +hf /xxh/xxh-dev/tests/xonsh/test_plugins.xsh ),
+        "1234\n5678\n\n{bar}\n{WHITE}{prompt_end}{NO_COLOR}"
+    )
+
 
     for shell in xxh_shell_repos.keys():
         for host, h in hosts.items():
@@ -154,60 +169,65 @@ if __name__ == '__main__':
                 continue
 
             print(f'\n[{shell} + {host}]')
-            user = h['user']
-            server = user + '@' + host
-            host_home = h['home']
 
-            check(
-                f'Remove {server}:~/.xxh',
-                $(echo @(h['sshpass']) ssh @(h['ssh_auth']) @(ssh_opts) @(server) "rm -rf ~/.xxh"),
-                ''
-            )
+            if host == 'local':
+                server = 'local'
+                host_home = h['home']
+            else:
+                user = h['user']
+                server = user + '@' + host
+                host_home = h['home']
 
-            check(
-                'Test connect using ssh',
-                $(echo @(h['sshpass']) ssh @(h['ssh_auth']) @(ssh_opts) @(server) "echo Test!"),
-                'Test!'
-            )
+                check(
+                    f'Remove {server}:~/.xxh',
+                    $(echo @(h['sshpass']) ssh @(h['ssh_auth']) @(ssh_opts) @(server) "rm -rf ~/.xxh"),
+                    ''
+                )
+
+                check(
+                    'Test connect using ssh',
+                    $(echo @(h['sshpass']) ssh @(h['ssh_auth']) @(ssh_opts) @(server) "echo Test!"),
+                    'Test!'
+                )
 
             if shell == 'xxh-shell-xonsh':
                 check(
                     'Test install xxh',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +iff +s @(shell) +hf /xxh/xxh-dev/tests/xonsh/test_env.py @(xxh_args) ),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +iff +s @(shell) +hf /xxh/xxh-dev/tests/xonsh/test_env.py ),
                     "{{'XXH_HOME': '{host_home}/.xxh', 'PYTHONPATH': '{host_home}/.xxh/.local/lib/python3.8/site-packages'}}".format(host_home=host_home)
                 )
 
                 # check(
                 #     'Test AppImage extraction on the host',
-                #     $(echo @(h['sshpass']) ssh @(h['ssh_auth']) @(ssh_opts) @(server) @(f"[ -d {host_home}/.xxh/.xxh/shells/xxh-shell-xonsh/build/xonsh-squashfs ] && echo 'extracted' ||echo 'not_extracted'") ),
+                #     $(echo @(h['sshpass']) ssh @(h['ssh_auth']) @(ssh_opts) @(server) @(xxh_args) @(f"[ -d {host_home}/.xxh/.xxh/shells/xxh-shell-xonsh/build/xonsh-squashfs ] && echo 'extracted' ||echo 'not_extracted'") ),
                 #     'not_extracted' if 'f' in host.split('_')[-1] else 'extracted'
                 # )
 
                 check(
                     'Test python',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +hf /xxh/xxh-dev/tests/xonsh/test_python.xsh @(xxh_args) ),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +hf /xxh/xxh-dev/tests/xonsh/test_python.xsh ),
                     "Python 3.8"
                 )
 
                 check(
                     'Test xonsh run xonsh script',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +hf /xxh/xxh-dev/tests/xonsh/test_xonsh_run_xonsh.xsh +e TESTENV="test env" @(xxh_args)),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +hf /xxh/xxh-dev/tests/xonsh/test_xonsh_run_xonsh.xsh +e TESTENV="test env" ),
                     "123\n.xxh\ntest env"
                 )
 
                 check(
                     'Test pip upgrade',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +hf /xxh/xxh-dev/tests/xonsh/test_pip_upgrade.xsh @(xxh_args)),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +hf /xxh/xxh-dev/tests/xonsh/test_pip_upgrade.xsh ),
                     ""
                 )
                 check(
                     'Test pip package install',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +hf /xxh/xxh-dev/tests/xonsh/test_pip_package_install.xsh @(xxh_args)),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +hf /xxh/xxh-dev/tests/xonsh/test_pip_package_install.xsh ),
                     ""
                 )
                 check(
                     'Test pip package import',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +hf /xxh/xxh-dev/tests/xonsh/test_pip_package_import.xsh @(xxh_args)),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +hf /xxh/xxh-dev/tests/xonsh/test_pip_package_import.xsh ),
                     "[[1], [2], [3]]"
                 )
 
@@ -215,13 +235,13 @@ if __name__ == '__main__':
 
                 check(
                     'Test xontrib',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +iff +hf /xxh/xxh-dev/tests/xonsh/test_xontrib.xsh @(xxh_args)),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +iff +hf /xxh/xxh-dev/tests/xonsh/test_xontrib.xsh ),
                     "autojump  installed      loaded\nschedule  installed      loaded"
                 )
 
                 check(
                     'Test xxh plugins',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +iff +hf /xxh/xxh-dev/tests/xonsh/test_plugins.xsh @(xxh_args)),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +iff +hf /xxh/xxh-dev/tests/xonsh/test_plugins.xsh ),
                     "1234\n5678\n\n{bar}\n{WHITE}{prompt_end}{NO_COLOR}"
                 )
 
@@ -229,33 +249,33 @@ if __name__ == '__main__':
                 shell_arg = ['+s', shell]
                 check(
                     'Test zsh env',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +if +hf /xxh/xxh-dev/tests/zsh/test_env.zsh +e TESTENV="test env" @(xxh_args) @(shell_arg) ),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +if +hf /xxh/xxh-dev/tests/zsh/test_env.zsh +e TESTENV="test env" @(shell_arg) ),
                     "test zsh .xxh and env=test env"
                 )
 
                 check(
                     'Test zsh command',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +if +hc @('"echo \'test \\"zsh\\" command\'"') @(xxh_args) @(shell_arg) ),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +if +hc @('"echo \'test \\"zsh\\" command\'"') @(xxh_args) ),
                     'test "zsh" command'
                 )
             elif shell == 'xxh-shell-bash':
                 shell_arg = ['+s', shell]
                 check(
                     'Test bash env',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +if +hf /xxh/xxh-dev/tests/bash/test_env.sh +e TESTENV="test env" @(xxh_args) @(shell_arg) ),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +if +hf /xxh/xxh-dev/tests/bash/test_env.sh +e TESTENV="test env" @(shell_arg) ),
                     "test bash xxh .xxh and env=test env"
                 )
 
                 check(
                     'Test bash command',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +if +hc @('"echo \'test \\"bash\\" command\'"') @(xxh_args) @(shell_arg) ),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +if +hc @('"echo \'test \\"bash\\" command\'"') @(shell_arg) ),
                     'test "bash" command'
                 )
 
                 # $OSH_THEME="morris"
                 # check(
                 #     'Test bash seamless',
-                #     $(echo source @(xxh)_xxh/xxh.bash @(h['xxh_auth']) @(server) +if +hf /xxh/xxh-dev/tests/bash/test_seamless.sh @(xxh_args) @(shell_arg) ),
+                #     $(echo source @(xxh)_xxh/xxh.bash @(h['xxh_auth']) @(server) @(xxh_args) +if +hf /xxh/xxh-dev/tests/bash/test_seamless.sh @(shell_arg) ),
                 #     "morris"
                 # )
 
@@ -263,7 +283,7 @@ if __name__ == '__main__':
                 shell_arg = ['+s', shell]
                 check(
                     'Test fish command',
-                    $(echo @(xxh) @(h['xxh_auth']) @(server) +if +hc @('"echo test"') @(xxh_args) @(shell_arg) ),
+                    $(echo @(xxh) @(h['xxh_auth']) @(server) @(xxh_args) +if +hc @('"echo test"') @(shell_arg) ),
                     "test"
                 )
